+++ src/main/java/com/hubspot/seatsolver/model/PointBase.java
@@ -1,6 +1,6 @@
 package com.hubspot.seatsolver.model;
 
 public interface PointBase {
-  int x();
-  int y();
+  double x();
+  double y();
 }
